fix: 데이터 타입 변경
fix: 데이터 타입 변경

+++ src/main/java/com/team/saver/market/basket/dto/BasketMenuResponse.java
@@ -19,6 +19,6 @@ public class BasketMenuResponse {
 
     private String optionDescription;
 
-    private long optionPrice;
+    private int optionPrice;
 
 }
